pyproject.toml
[project]
name = "doc-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.13.1",
    "fastapi>=0.109.0",
    "openai>=1.12.0",
    "passlib[bcrypt]>=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=10.2.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "streamlit>=1.31.0",
    "typing-extensions>=4.9.0",
    "uvicorn>=0.27.0",
]
